+++ src/main/java/com/sistemaguincho/gestaoguincho/dto/ValoresFixosRequest.java
@@ -2,14 +2,16 @@ package com.sistemaguincho.gestaoguincho.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ValoresFixosRequest {
 
-    private Double valorQuilometragemPorKm;
-    private Double valorQuilometragemSaida;
-    private Double valorMotoristaPorKm;
-    private Double valorMotoristaSaida;
-    private Double valorHoraParada;
-    private Double valorHoraTrabalhada;
-    private Double valorDiaria;
+    private BigDecimal valorQuilometragemPorKm;
+    private BigDecimal  valorQuilometragemSaida;
+    private BigDecimal  valorMotoristaPorKm;
+    private BigDecimal  valorMotoristaSaida;
+    private BigDecimal  valorHoraParada;
+    private BigDecimal  valorHoraTrabalhada;
+    private BigDecimal  valorDiaria;
 }
